fix(tasks): harden lease gate and seams
