+++ packages/api/Dockerfile
@@ -1,13 +1,20 @@
-# openagent.email API — built and run with Bun (no transpile step; src is TS).
-FROM oven/bun:1 AS deps
+# openagent.email API — build the canonical dist artifact, then ship only it.
+FROM oven/bun:1 AS production-deps
 WORKDIR /app
 COPY package.json bun.lock* ./
 RUN bun install --production --frozen-lockfile || bun install --production
 
-FROM oven/bun:1
+FROM oven/bun:1 AS build
+WORKDIR /app
+COPY package.json bun.lock* ./
+RUN bun install --frozen-lockfile
+COPY . .
+RUN bun run build
+
+FROM oven/bun:1 AS runtime
 WORKDIR /app
 ENV NODE_ENV=production
-COPY --from=deps /app/node_modules ./node_modules
-COPY . .
+COPY --from=production-deps /app/node_modules ./node_modules
+COPY --from=build /app/dist ./dist
 EXPOSE 3100
-CMD ["bun", "run", "src/main.ts"]
+CMD ["bun", "dist/main.js"]
