+++ packages/api/test/mcp-http.test.ts
@@ -630,7 +630,6 @@ describe('MCP registered task handlers execute through the production HTTP trans
     const unused = async () => { throw new Error('unused in R9 lease fixture'); };
     const handlerApp = createApp({
       uiEnabled: false,
-      leaseEnabledForTests: true,
       taskService: {
         create: unused, list: unused, listBoard: unused, get: async () => base,
         update: unused, reply: unused, remind: unused, close: unused, waitForTerminal: unused,
