+++ packages/api/test/support/task-lease-seams.ts
@@ -1,0 +1,15 @@
+import {
+  claimLeaseHeadersForTests,
+  parseTaskMessageForTests,
+  parseStampedTaskMessageForTests,
+  setTaskLeasesEnabledForTests,
+  type RawTaskMessage,
+  type TaskState,
+} from '../../src/lib/tasks-internal.ts';
+
+export {
+  claimLeaseHeadersForTests,
+  parseTaskMessageForTests,
+  parseStampedTaskMessageForTests,
+  setTaskLeasesEnabledForTests,
+};
