+++ packages/api/test/task-lease-core.test.ts
@@ -15,6 +15,7 @@ process.env.SMTP_USER = 'agent@test.example';
 process.env.SMTP_PASS = 'smtp-secret';
 process.env.DATA_DIR = mkdtempSync(join(tmpdir(), 'oae-task-lease-core-'));
 process.env.TASK_LEASES_ENABLED = 'true';
+process.env.NODE_ENV = 'test';
 
 const { afterEach, describe, expect, test } = await import('bun:test');
 const { Hono } = await import('hono');
@@ -24,21 +25,19 @@ const {
   TASK_LEASE_MAX_SEC,
   TASK_LEASE_MIN_SEC,
   TASK_LEASE_REASON_MAX_CHARS,
-  claimLeaseHeadersForTests,
   claimTask,
   createApprovalTask,
   clearQueuedEventsForTests,
   isTaskLeaseTokenCurrent,
-  parseTaskMessageForTests,
   releaseTask,
   setTaskGetForTests,
-  setTaskLeasesEnabledForTests,
   setTaskNowForTests,
   setTaskSendMailForTests,
   taskService,
   taskFromMessages,
   toTaskView,
 } = await import('../src/lib/tasks.ts');
+const { claimLeaseHeadersForTests, parseTaskMessageForTests, setTaskLeasesEnabledForTests } = await import('./support/task-lease-seams.ts');
 const { createIdentity } = await import('../src/lib/identities.ts');
 const { createTaskRoutes } = await import('../src/routes/tasks.ts');
 
@@ -179,7 +178,8 @@ describe('#56 R2 lease authority', () => {
     };
     const app = new Hono();
     app.use('*', async (c, next) => { c.set('auth', { kind: 'identity' as const, address: B }); await next(); });
-    app.route('/v1/tasks', createTaskRoutes({ service, findIdentity: () => ({ address: B, createdAt: '2026-08-24T00:00:00.000Z' }), leaseEnabledForTests: false }));
+    setTaskLeasesEnabledForTests(false);
+    app.route('/v1/tasks', createTaskRoutes({ service, findIdentity: () => ({ address: B, createdAt: '2026-08-24T00:00:00.000Z' }) }));
     const response = await app.request(`/v1/tasks/${ID}/claim`, {
       method: 'POST', headers: { 'content-type': 'application/json' }, body: JSON.stringify({ leaseSec: 300 }),
     });
@@ -306,7 +306,6 @@ describe('#56 R2 lease authority', () => {
     app.use('*', async (c, next) => { c.set('auth', { kind: 'identity' as const, address: B }); await next(); });
     app.route('/v1/tasks', createTaskRoutes({
       findIdentity: (address) => ({ address, createdAt: '2026-08-24T00:00:00.000Z' }),
-      leaseEnabledForTests: true,
     }));
     const response = await app.request(`/v1/tasks/${ID}/claim`, {
       method: 'POST', headers: { 'content-type': 'application/json' }, body: JSON.stringify({ leaseSec: 300 }),
@@ -374,7 +373,6 @@ function productionApp(service: TaskService = taskService, actor = B) {
   app.route('/v1/tasks', createTaskRoutes({
     service,
     findIdentity: (address) => ({ address, createdAt: '2026-08-24T00:00:00.000Z' }),
-    leaseEnabledForTests: true,
   }));
   return app;
 }
