+++ packages/api/test/task-lease-reaper-red.test.ts
@@ -14,13 +14,12 @@ process.env.SMTP_USER = 'agent@test.example';
 process.env.SMTP_PASS = 'smtp-secret';
 process.env.DATA_DIR = mkdtempSync(join(tmpdir(), 'oae-task-lease-reaper-red-'));
 process.env.TASK_LEASES_ENABLED = 'true';
+process.env.NODE_ENV = 'test';
 
 const { afterEach, describe, expect, test } = await import('bun:test');
 const {
   clearQueuedEventsForTests,
-  claimLeaseHeadersForTests,
   isTaskLeaseTokenCurrent,
-  parseTaskMessageForTests,
   reapExpiredTaskLeasesOnce,
   setTaskGetForTests,
   setTaskListAllForTests,
@@ -30,6 +29,7 @@ const {
   taskService,
   toTaskView,
 } = await import('../src/lib/tasks.ts');
+const { claimLeaseHeadersForTests, parseTaskMessageForTests } = await import('./support/task-lease-seams.ts');
 const { startTaskLeaseReaper, TASK_LEASE_REAPER_INTERVAL_MS } = await import('../src/lib/task-lease-reaper.ts');
 
 const ID = '0fdc3207-056e-47c1-a65c-b29d39f66b83';
@@ -752,7 +752,6 @@ describe('#56 R8b explicit server lease expiry reaper RED', () => {
     let release: (() => void) | undefined;
     const pending = new Promise<void>((resolve) => { release = resolve; });
     startTaskLeaseReaper({
-      leaseEnabledForTests: true,
       reapOnce: async () => {
         calls += 1;
         if (calls === 1) await pending;
