+++ packages/api/test/task-lease-seams.test.ts
@@ -1,0 +1,52 @@
+import { readFileSync } from 'node:fs';
+import { join } from 'node:path';
+
+process.env.DOMAIN = 'test.example';
+process.env.API_KEYS = 'admin-key';
+process.env.IMAP_USER = 'agent@test.example';
+process.env.IMAP_PASS = 'imap-secret';
+process.env.SMTP_USER = 'agent@test.example';
+process.env.SMTP_PASS = 'smtp-secret';
+process.env.TASK_LEASES_ENABLED = 'true';
+
+const { expect, test } = await import('bun:test');
+
+test('#89 RED: production lease modules expose neither signing/parser helpers nor gate overrides', async () => {
+  const tasks = await import('../src/lib/tasks.ts');
+  expect(Object.hasOwn(tasks, 'claimLeaseHeadersForTests')).toBe(false);
+  expect(Object.hasOwn(tasks, 'parseTaskMessageForTests')).toBe(false);
+  expect(Object.hasOwn(tasks, 'setTaskLeasesEnabledForTests')).toBe(false);
+  expect(Object.hasOwn(tasks, 'parseStampedTaskMessageForTests')).toBe(false);
+  for (const path of ['../src/routes/tasks.ts', '../src/lib/task-lease-reaper.ts', '../src/app.ts']) {
+    expect(readFileSync(new URL(path, import.meta.url), 'utf8')).not.toContain('leaseEnabledForTests');
+  }
+});
+
+test('#89 GREEN: canonical production bundle and final Docker stage exclude every lease test seam', () => {
+  const pkgDir = join(import.meta.dir, '..');
+  const build = Bun.spawnSync(['bun', 'run', 'build'], { cwd: pkgDir });
+  expect(build.exitCode).toBe(0);
+  const bundle = readFileSync(join(pkgDir, 'dist', 'main.js'), 'utf8');
+  for (const seam of [
+    'claimLeaseHeadersForTests',
+    'parseTaskMessageForTests',
+    'setTaskLeasesEnabledForTests',
+    'parseStampedTaskMessageForTests',
+    'taskLeaseTestRegistry',
+  ]) expect(bundle).not.toContain(seam);
+
+  const runtimeStage = readFileSync(join(pkgDir, 'Dockerfile'), 'utf8').split('FROM oven/bun:1 AS runtime', 2)[1]!;
+  const runtimeCopies = [...runtimeStage.matchAll(/^COPY\s+--from=(\S+)\s+(\S+)\s+(\S+)\s*$/gm)]
+    .map(([, stage, source, destination]) => [stage, source, destination]);
+  // This is an image-stage proof independent of an installed Docker daemon:
+  // the final stage has exactly these two artifacts, never the build context.
+  expect(runtimeCopies).toEqual([
+    ['production-deps', '/app/node_modules', './node_modules'],
+    ['build', '/app/dist', './dist'],
+  ]);
+  expect(runtimeStage).not.toMatch(/^ADD\s+/m);
+  expect(runtimeStage).not.toContain('COPY . .');
+  expect(runtimeStage).not.toContain('src/');
+  expect(runtimeStage).not.toContain('test/');
+  expect(runtimeStage).toContain('CMD ["bun", "dist/main.js"]');
+});
