+++ packages/api/test/task-leases-red.test.ts
@@ -12,10 +12,12 @@ process.env.IMAP_PASS = 'imap-secret';
 process.env.SMTP_USER = 'agent@test.example';
 process.env.SMTP_PASS = 'smtp-secret';
 process.env.DATA_DIR = mkdtempSync(join(tmpdir(), 'oae-task-leases-red-'));
+process.env.NODE_ENV = 'test';
 
 const { afterEach, beforeEach, describe, expect, test } = await import('bun:test');
 const { Hono } = await import('hono');
 const { setTaskNowForTests } = await import('../src/lib/tasks.ts');
+const { setTaskLeasesEnabledForTests } = await import('./support/task-lease-seams.ts');
 const { createTaskRoutes } = await import('../src/routes/tasks.ts');
 
 const ID = '0fdc3207-056e-47c1-a65c-b29d39f66b83';
@@ -111,6 +113,7 @@ const capableService = {
 } as LeaseCapableService;
 
 function appFor(address: string, leaseEnabledForTests = true, selectedService: TaskService = capableService) {
+  setTaskLeasesEnabledForTests(leaseEnabledForTests);
   const app = new Hono();
   app.use('*', async (c, next) => {
     c.set('auth', { kind: 'identity' as const, address });
@@ -121,7 +124,6 @@ function appFor(address: string, leaseEnabledForTests = true, selectedService: T
     findIdentity: (candidate) => [REQUESTER, RECIPIENT, OUTSIDER].includes(candidate.toLowerCase())
       ? { address: candidate, createdAt: '2026-08-24T00:00:00.000Z' }
       : undefined,
-    leaseEnabledForTests,
   }));
   return app;
 }
@@ -136,6 +138,7 @@ beforeEach(() => {
 });
 
 afterEach(() => {
+  setTaskLeasesEnabledForTests(undefined);
   setTaskNowForTests(null);
 });
